requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
black==25.1.0
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   pip-tools
distlib==0.3.9
    # via virtualenv
duckdb==1.2.2
    # via -r requirements.in
filelock==3.18.0
    # via virtualenv
flake8==7.2.0
    # via
    #   -r requirements.in
    #   flake8-docstrings
flake8-docstrings==1.7.0
    # via -r requirements.in
google-api-core==2.24.2
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.1.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.69.2
    # via google-api-core
identify==2.6.9
    # via pre-commit
idna==3.10
    # via requests
mccabe==0.7.0
    # via flake8
mypy==1.15.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nba-api==1.9.0
    # via -r requirements.in
nodeenv==1.9.1
    # via pre-commit
numpy==2.2.4
    # via
    #   nba-api
    #   pandas
    #   pandas-stubs
packaging==24.2
    # via
    #   black
    #   build
    #   wheel
pandas==2.2.3
    # via nba-api
pandas-stubs==2.2.3.250308
    # via -r requirements.in
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.7
    # via
    #   black
    #   virtualenv
pre-commit==4.2.0
    # via -r requirements.in
proto-plus==1.26.1
    # via google-api-core
protobuf==4.25.6
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyarrow==19.0.1
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycodestyle==2.13.0
    # via flake8
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==3.3.2
    # via flake8
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via pre-commit
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
    #   nba-api
rsa==4.9
    # via google-auth
six==1.17.0
    # via python-dateutil
snowballstemmer==2.2.0
    # via pydocstyle
tabulate==0.9.0
    # via -r requirements.in
types-pytz==2025.2.0.20250326
    # via
    #   -r requirements.in
    #   pandas-stubs
types-requests==2.32.0.20250328
    # via -r requirements.in
types-tabulate==0.9.0.20241207
    # via -r requirements.in
typing-extensions==4.13.1
    # via mypy
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via
    #   requests
    #   types-requests
virtualenv==20.30.0
    # via pre-commit
wheel==0.46.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
